SExpressions.jl

SExpressions.jl includes a roughly Racket-compatible parser for s-expressions.

Core Parser Interface

SExpressions.jl

Core Parser Interface

SExpressions.Parser.parse — Function.

Parse a Symbol or number or Keyword or Bool at the given index.

source

Parse a String at the given index.

source

Parse a list at the given index.

source

Parse an s-expression at the given index of the given AbstractString.

source

SExpressions.Parser.parses — Function.

Parse an entire string into a single list.

source

SExpressions.Parser.parsefile — Function.

Parse a file into a single list.

source