SExpressions.jl

SExpressions.jl includes a roughly Racket-compatible parser for s-expressions.

SExpressions.jl

Lists

Core Parser Interface

Lists

Core Parser Interface

SExpressions.Parser.parse — Function

parse(s::AbstractString)

Read the given string s as a single s-expression.

parse(io::IO)

Read a single object from the given io stream.

SExpressions.Parser.parseall — Function

parseall(io::IO)
parseall(s::AbstractString)

Parse all objects from the given stream or string into a single list.

SExpressions.Parser.parsefile — Function

parsefile(filename::AbstractString)

Parse a file into a single list.